+++ src/utils.js
@@ -3,8 +3,7 @@ const { ServerStyleSheet, isStyledComponent } = require('styled-components')
 
 let StyleSheet
 
-/* eslint-disable */
-if (!!isStyledComponent) {
+if (isStyledComponent) {
   const secretInternals = require('styled-components')
     .__DO_NOT_USE_OR_YOU_WILL_BE_HAUNTED_BY_SPOOKY_GHOSTS
 
@@ -21,7 +20,6 @@ if (!!isStyledComponent) {
 } else {
   StyleSheet = require('styled-components/lib/models/StyleSheet').default
 }
-/* eslint-enable */
 
 const isServer = () => typeof document === 'undefined'
 
@@ -58,8 +56,33 @@ const getHashes = () =>
     .concat(getComponentIDs())
     .filter(Boolean)
 
+const buildReturnMessage = (utils, pass, property, received, expected) => () =>
+  `${utils.printReceived(
+    !received && !pass
+      ? `Property '${property}' not found in style rules`
+      : `Value mismatch for property '${property}'`
+  )}\n\n` +
+  'Expected\n' +
+  `  ${utils.printExpected(`${property}: ${expected}`)}\n` +
+  'Received:\n' +
+  `  ${utils.printReceived(`${property}: ${received}`)}`
+
+const matcherTest = (received, expected) => {
+  try {
+    const matcher =
+      expected instanceof RegExp ? expect.stringMatching(expected) : expected
+
+    expect(received).toEqual(matcher)
+    return true
+  } catch (error) {
+    return false
+  }
+}
+
 module.exports = {
   resetStyleSheet,
   getCSS,
   getHashes,
+  buildReturnMessage,
+  matcherTest,
 }